Add OpenChannelModal
