+++ src/sections/channels.tsx
@@ -14,7 +14,7 @@ import {
   DialogContent,
   TextField,
   DialogActions,
-  InputAdornment
+  InputAdornment,
 } from '@mui/material';
 import Section from '../future-hopr-lib-components/Section';
 import { useAppDispatch, useAppSelector } from '../store';
@@ -25,6 +25,7 @@ import { exportToCsv } from '../utils/helpers';
 import CircularProgress from '@mui/material/CircularProgress';
 import { FundChannelModal } from '../components/FundChannelModal';
 import { ethers } from 'ethers';
+import { OpenChannelModal } from '../components/OpenChannelModal';
 
 function ChannelsPage() {
   const dispatch = useAppDispatch();
@@ -45,100 +46,10 @@ function ChannelsPage() {
       }
     >
   >({});
-  const [peerId, set_peerId] = useState('');
-  const [amount, set_amount] = useState('');
-  const [openChannelDialog, set_openChannelDialog] = useState(false);
-  const [channelOpening, set_channelOpening] = useState(false);
-  const [openingErrors, set_openingErrors] = useState<{ status: string | undefined; error: string | undefined }[]>([]);
-  const [openingSuccess, set_openingSucess] = useState(false);
+
   const [queryParams, set_queryParams] = useState('');
 
   const navigate = useNavigate();
-
-  const handleOpenChannelDialog = () => {
-    set_openChannelDialog(true);
-  };
-
-  const handleCloseChannelDialog = () => {
-    set_openChannelDialog(false);
-  };
-
-  const openChannelPopUp = () => {
-    return (
-      <>
-        <button onClick={handleOpenChannelDialog}>Open Channel</button>
-        <Dialog
-          open={openChannelDialog}
-          onClose={handleCloseChannelDialog}
-        >
-          <DialogTitle>Open Channel</DialogTitle>
-          <DialogContent>
-            <TextField
-              label="Peer ID"
-              value={peerId}
-              placeholder="16Eiu2HAm..."
-              onChange={(e) => set_peerId(e.target.value)}
-            />
-            <TextField
-              label="Amount"
-              type="string"
-              value={amount}
-              onChange={(e) => set_amount(e.target.value)}
-              InputProps={{ endAdornment: <InputAdornment position="end">mHOPR</InputAdornment> }}
-            />
-          </DialogContent>
-          <DialogActions>
-            <button onClick={handleCloseChannelDialog}>Cancel</button>
-            <button
-              onClick={() => handleOpenChannel(amount, peerId)}
-              disabled={!amount || parseFloat(amount) <= 0 || !peerId}
-            >
-              Open Channel
-            </button>
-          </DialogActions>
-        </Dialog>
-        {channelOpening && <CircularProgress />}
-        {openingSuccess && <div>Opening Channel Success</div>}
-        {openingErrors.map((error, index) => (
-          <div key={index}>{error.error}</div>
-        ))}
-      </>
-    );
-  };
-
-  const handleOpenChannel = (amount: string, peerId: string) => {
-    set_channelOpening(true); // Set loading state
-
-    dispatch(
-      actionsAsync.openChannelThunk({
-        apiEndpoint: loginData.apiEndpoint!,
-        apiToken: loginData.apiToken!,
-        amount: amount,
-        peerId: peerId,
-        timeout: 60e3,
-      }),
-    )
-      .unwrap()
-      .then(() => {
-        // handle success opening channel
-        set_channelOpening(false);
-        set_openingSucess(true);
-        set_openingErrors([]);
-        handleRefresh();
-      })
-      .catch((e) => {
-        set_openingSucess(false);
-        set_openingErrors([
-          ...openingErrors,
-          {
-            error: e.error,
-            status: e.status,
-          },
-        ]);
-        set_channelOpening(false);
-        //handle error on opening channel});
-      });
-  };
 
   const handleTabChange = (event: React.SyntheticEvent<Element, Event>, newTabIndex: number) => {
     set_tabIndex(newTabIndex);
@@ -176,13 +87,13 @@ function ChannelsPage() {
       actionsAsync.getChannelsThunk({
         apiEndpoint: loginData.apiEndpoint!,
         apiToken: loginData.apiToken!,
-      }),
+      })
     );
     dispatch(
       actionsAsync.getAliasesThunk({
         apiEndpoint: loginData.apiEndpoint!,
         apiToken: loginData.apiToken!,
-      }),
+      })
     );
   };
 
@@ -213,7 +124,7 @@ function ChannelsPage() {
                 status: channel.status,
                 dedicatedFunds: channel.balance,
               })),
-              `${tabLabel}-channels.csv`,
+              `${tabLabel}-channels.csv`
             );
           }
         }}
@@ -239,7 +150,7 @@ function ChannelsPage() {
         apiToken: loginData.apiToken!,
         direction: direction,
         peerId: peerId,
-      }),
+      })
     )
       .unwrap()
       .then(() => {
@@ -292,7 +203,7 @@ function ChannelsPage() {
         </Tabs>
       </Box>
       {exportToCsvButton()}
-      {tabIndex === 1 && openChannelPopUp()}
+      {tabIndex === 1 && <OpenChannelModal handleRefresh={handleRefresh} />}
       <TableContainer component={Paper}>
         <Table
           sx={{ minWidth: 650 }}
