+++ src/store/slices/node/actionsAsync.ts
@@ -19,7 +19,7 @@ import {
   type SignPayloadType,
   type WithdrawPayloadType,
   api,
-  utils,
+  utils
 } from '@hoprnet/hopr-sdk';
 
 const { APIError } = utils;
@@ -155,7 +155,7 @@ const getPeerInfoThunk = createAsyncThunk(
         });
       }
     }
-  }
+  },
 );
 
 const getSettingsThunk = createAsyncThunk('node/getSettings', async (payload: BasePayloadType, { rejectWithValue }) => {
@@ -186,7 +186,7 @@ const getStatisticsThunk = createAsyncThunk(
         });
       }
     }
-  }
+  },
 );
 
 const getTicketsThunk = createAsyncThunk('node/getTickets', async (payload: BasePayloadType, { rejectWithValue }) => {
@@ -231,7 +231,7 @@ const getEntryNodesThunk = createAsyncThunk(
         });
       }
     }
-  }
+  },
 );
 
 const getVersionThunk = createAsyncThunk('node/getVersion', async (payload: BasePayloadType, { rejectWithValue }) => {
@@ -314,7 +314,7 @@ const removeAliasThunk = createAsyncThunk(
         });
       }
     }
-  }
+  },
 );
 
 const closeChannelThunk = createAsyncThunk(
@@ -331,7 +331,7 @@ const closeChannelThunk = createAsyncThunk(
         });
       }
     }
-  }
+  },
 );
 
 const fundChannelsThunk = createAsyncThunk(
@@ -348,7 +348,7 @@ const fundChannelsThunk = createAsyncThunk(
         });
       }
     }
-  }
+  },
 );
 
 const getChannelThunk = createAsyncThunk(
@@ -365,7 +365,7 @@ const getChannelThunk = createAsyncThunk(
         });
       }
     }
-  }
+  },
 );
 
 const getChannelTicketsThunk = createAsyncThunk(
@@ -382,7 +382,7 @@ const getChannelTicketsThunk = createAsyncThunk(
         });
       }
     }
-  }
+  },
 );
 
 const openChannelThunk = createAsyncThunk(
@@ -399,7 +399,7 @@ const openChannelThunk = createAsyncThunk(
         });
       }
     }
-  }
+  },
 );
 
 const redeemChannelTicketsThunk = createAsyncThunk(
@@ -416,7 +416,7 @@ const redeemChannelTicketsThunk = createAsyncThunk(
         });
       }
     }
-  }
+  },
 );
 
 const sendMessageThunk = createAsyncThunk(
@@ -436,7 +436,7 @@ const sendMessageThunk = createAsyncThunk(
         });
       }
     }
-  }
+  },
 );
 
 const signThunk = createAsyncThunk('node/sign', async (payload: SignPayloadType, { rejectWithValue }) => {
@@ -484,7 +484,7 @@ const setSettingThunk = createAsyncThunk(
         });
       }
     }
-  }
+  },
 );
 
 const redeemTicketsThunk = createAsyncThunk(
@@ -501,7 +501,7 @@ const redeemTicketsThunk = createAsyncThunk(
         });
       }
     }
-  }
+  },
 );
 
 const createTokenThunk = createAsyncThunk(
@@ -518,7 +518,7 @@ const createTokenThunk = createAsyncThunk(
         });
       }
     }
-  }
+  },
 );
 
 const deleteTokenThunk = createAsyncThunk(
@@ -538,7 +538,7 @@ const deleteTokenThunk = createAsyncThunk(
         });
       }
     }
-  }
+  },
 );
 
 const getPrometheusMetricsThunk = createAsyncThunk(
@@ -555,7 +555,7 @@ const getPrometheusMetricsThunk = createAsyncThunk(
         });
       }
     }
-  }
+  },
 );
 
 export const createExtraReducers = (builder: ActionReducerMapBuilder<typeof initialState>) => {
@@ -662,7 +662,13 @@ export const createExtraReducers = (builder: ActionReducerMapBuilder<typeof init
   });
   builder.addCase(getChannelThunk.fulfilled, (state, action) => {
     if (action.payload) {
-      const { balance, channelId, peerId, status, type } = action.payload;
+      const {
+        balance,
+        channelId,
+        peerId,
+        status,
+        type,
+      } = action.payload;
       // find channel if it already exists
       const channelIndex = state.channels?.[type].findIndex((channel) => channel.channelId === channelId);
